deletes extra folders and cleaned up current folders and adds spec folders
