--- lib/season.rb
@@ -1,5 +1,0 @@
-require 'csv'
-
-class Season
-    
-end
