+++ spec/league_stat_spec.rb
